Moving ColumnFamily and ColumnFamilyDescriptor to separate module...
